+++ actix-files/tests/encoding.rs
@@ -35,4 +35,12 @@ async fn test_utf8_file_contents() {
         res.headers().get(header::CONTENT_TYPE),
         Some(&HeaderValue::from_static("text/plain; charset=utf-8")),
     );
+
+    // prevent directory traversal attack
+    let srv = test::init_service(App::new().service(Files::new("/", "./tests"))).await;
+
+    let req = TestRequest::with_uri("..%5c/README.md").to_request();
+    let res = test::call_service(&srv, req).await;
+
+    assert_eq!(res.status(), StatusCode::FORBIDDEN);
 }